+++ Android/UdacityExercises/Lesson09b-ToDo-List-AAC/T09b.05-Exercise-DeleteTask/app/src/main/java/com/example/android/todolist/TaskAdapter.java
@@ -30,6 +30,7 @@ import com.example.android.todolist.database.TaskEntry;
 import java.text.SimpleDateFormat;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 
 /**
  * This TaskAdapter creates and binds ViewHolders, that hold the description and priority of a task,
@@ -135,7 +136,9 @@ public class TaskAdapter extends RecyclerView.Adapter<TaskAdapter.TaskViewHolder
         return mTaskEntries.size();
     }
 
-    // TODO (2) Add a getTasks method that returns mTaskEntries
+    public List<TaskEntry> getTasks() {
+        return mTaskEntries;
+    }
 
     /**
      * When data changes, this method updates the list of taskEntries